rename collateral2compiler folder dsl
