--- lake/collateral2compiler/lake_imports.py
@@ -1,3 +1,0 @@
-from lake.collateral2compiler.top_lake import TopLake
-from lake.collateral2compiler.mem_port import MemPort
-from lake.collateral2compiler.helper import make_params
