+++ lake/dsl/dsl_examples/memtile.py
@@ -1,4 +1,4 @@
-from lake.collateral2compiler.lake_imports import *
+from lake.dsl.lake_imports import *
 
 # example of DSL (makes current mem tile with 2 agg,
 # wide SRAM, 2 tb
