+++ lake/dsl/lake_imports.py
@@ -1,0 +1,3 @@
+from lake.dsl.top_lake import TopLake
+from lake.dsl.mem_port import MemPort
+from lake.dsl.helper import make_params
